+++ Modules/Intent.Modules.Modelers.Services.EventInteractions/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 1.1.0
 
 - Improvement: Updated Module from .NET Core 3.5 to .NET 8.
+- Improvement: Added extension methods for getting messages, message DTOs and message Enums based on being used in an association of any type the services designer.
 
 ### Version 1.0.4